fix(lint): resolve 10 golangci-lint failures (gs-50i)
- Replace crypto/sha1 with hash/fnv in classifierFingerprint (G505)
- Add G706 exclusion for internal/ taint analysis false positives
- Fix 5 British-English misspellings (recognised→recognized, etc.)
- Remove unnecessary uint64() conversions in tmp_inodes_check_unix.go
- Remove unused reason param from isPlannedPolecatExit + update tests

+++ internal/doctor/tmp_inodes_check_unix.go
@@ -16,8 +16,8 @@ func readTmpInodeUsage(path string) (tmpInodeUsage, error) {
 		return tmpInodeUsage{}, fmt.Errorf("statfs %s: %w", path, err)
 	}
 
-	total := uint64(st.Files)
-	free := uint64(st.Ffree)
+	total := st.Files
+	free := st.Ffree
 	var used uint64
 	if total >= free {
 		used = total - free
